+++ .eslintrc.js
@@ -36,7 +36,7 @@ module.exports = {
     "semi": ["error", "always"],
 
     // @typescript-eslint (https://github.com/typescript-eslint/typescript-eslint#readme)
-    "@typescript-eslint/no-unused-vars": ["warn", { argsIgnorePattern: "_+" }],
+    "@typescript-eslint/no-unused-vars": ["error", { argsIgnorePattern: "_+" }],
 
     // eslint-plugin-jest (https://github.com/jest-community/eslint-plugin-jest#readme)
     "jest/consistent-test-it": ["error", {