+++ test/main.test.ts
@@ -1,3 +1,5 @@
+import { format as strFormat } from "util";
+
 import * as core from "./mocks/@actions/core.mock";
 import * as github from "./mocks/@actions/github.mock";
 import { PR_NUMBER } from "./mocks/@actions/github.mock";
@@ -15,6 +17,7 @@ jest.mock("../src/svgo", () => svgo);
 
 import contentPayloads from "./fixtures/contents-payloads.json";
 import files from "./fixtures/file-data.json";
+import actionOptions from "./fixtures/svgo-action.json";
 import svgoOptions from "./fixtures/svgo-options.json";
 
 import { PR_NOT_FOUND } from "../src/github-api";
@@ -175,7 +178,12 @@ describe("Scenarios", () => {
     expect(githubAPI.commitFiles).toHaveBeenCalledWith(
       github.GitHubInstance,
       expect.any(Object),
-      expect.any(String),
+      expect.stringContaining("1"),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${testFilePath}`),
     );
   });
 
@@ -205,7 +213,12 @@ describe("Scenarios", () => {
     expect(githubAPI.commitFiles).toHaveBeenCalledWith(
       github.GitHubInstance,
       expect.any(Object),
-      expect.any(String),
+      expect.stringContaining("1"),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${fooFilePath}`),
     );
   });
 
@@ -256,7 +269,17 @@ describe("Scenarios", () => {
     expect(githubAPI.commitFiles).toHaveBeenCalledWith(
       github.GitHubInstance,
       expect.any(Object),
-      expect.any(String),
+      expect.stringContaining("2"),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${fooFilePath}`),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${barFilePath}`),
     );
   });
 
@@ -322,7 +345,12 @@ describe("Scenarios", () => {
     expect(githubAPI.commitFiles).toHaveBeenCalledWith(
       github.GitHubInstance,
       expect.any(Object),
-      expect.any(String),
+      expect.stringContaining("1"),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${testFilePath}`),
     );
   });
 
@@ -352,7 +380,12 @@ describe("Scenarios", () => {
     expect(githubAPI.commitFiles).toHaveBeenCalledWith(
       github.GitHubInstance,
       expect.any(Object),
-      expect.any(String),
+      expect.stringContaining("1"),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${testFilePath}`),
     );
   });
 
@@ -382,7 +415,12 @@ describe("Scenarios", () => {
     expect(githubAPI.commitFiles).toHaveBeenCalledWith(
       github.GitHubInstance,
       expect.any(Object),
-      expect.any(String),
+      expect.stringContaining("1"),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${complexFilePath}`),
     );
   });
 
@@ -437,7 +475,22 @@ describe("Scenarios", () => {
     expect(githubAPI.commitFiles).toHaveBeenCalledWith(
       github.GitHubInstance,
       expect.any(Object),
-      expect.any(String),
+      expect.stringContaining("3"),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${fooFilePath}`),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${barFilePath}`),
+    );
+    expect(githubAPI.commitFiles).toHaveBeenCalledWith(
+      github.GitHubInstance,
+      expect.any(Object),
+      expect.stringContaining(`- ${testFilePath}`),
     );
   });
 
@@ -450,7 +503,7 @@ describe("Scenarios", () => {
     expect(core.debug).toHaveBeenCalledWith(expect.stringMatching(/skipping.*optimized.svg/));
   });
 
-  test("use a configuration file in the repository", async () => {
+  test("use a SVGO configuration file in the repository", async () => {
     svgo.getDefaultSvgoOptions.mockReturnValueOnce(svgoOptions);
 
     await main();
@@ -463,13 +516,41 @@ describe("Scenarios", () => {
     PR_NUMBER.MODIFY_SVG,
     PR_NUMBER.MANY_CHANGES,
   ])("dry run enabled (#%i)", async (prNumber) => {
-    inputs.getDryRun.mockReturnValueOnce(true);
+    inputs.ActionConfigInstance.isDryRun.mockReturnValue(true);
     githubAPI.getPrNumber.mockReturnValueOnce(prNumber);
 
     await main();
+    inputs.ActionConfigInstance.isDryRun.mockReturnValue(false);
 
     expect(githubAPI.commitFiles).not.toHaveBeenCalled();
     expect(core.info).toHaveBeenCalledWith(expect.stringContaining("Dry mode enabled"));
+  });
+
+  test.each([
+    "This is the commit title\n\nAnd this the message (%s)",
+    "chore: make some changes\n\n- This isn't tennis\n- Praise the sun\n\n%s",
+    "Added some SVGs to the website\n\n%s",
+    "Double rainbow\n\nwhat does it %s mean?",
+  ])("disabled from commit message", async (baseCommitMessage) => {
+    const fullCommitMessage = strFormat(baseCommitMessage, "disable-svgo-action");
+    githubAPI.getCommitMessage.mockResolvedValueOnce(fullCommitMessage);
+
+    const result = await main();
+
+    expect(result).toBe(true);
+    expect(githubAPI.commitFiles).not.toHaveBeenCalled();
+    expect(core.info).toHaveBeenCalledWith(expect.stringContaining("disabled"));
+  });
+
+  test("custom configuration file usage", async () => {
+    const actionConfigFilePath = "svgo-action.yml";
+    inputs.getConfigFilePath.mockReturnValueOnce(actionConfigFilePath);
+
+    await main();
+
+    const { content, encoding } = contentPayloads[actionConfigFilePath];
+    expect(encoder.decode).toHaveBeenCalledWith(content, encoding);
+    expect(inputs.ActionConfig).toHaveBeenCalledWith(actionOptions);
   });
 
 });
@@ -493,9 +574,19 @@ describe("Error scenarios", () => {
     expect(core.setFailed).toHaveBeenCalledTimes(1);
   });
 
-  test("there is no configuration file in repository", async () => {
+  test("there is no SVGO configuration file in repository", async () => {
     githubAPI.getPrNumber.mockReturnValueOnce(PR_NUMBER.ADD_SVG);
     svgo.getDefaultSvgoOptions.mockResolvedValueOnce({ });
+
+    await main();
+
+    expect(core.setFailed).not.toHaveBeenCalled();
+    expect(githubAPI.commitFiles).toHaveBeenCalledTimes(1);
+  });
+
+  test("custom configuration file does not exist", async () => {
+    githubAPI.getPrNumber.mockReturnValueOnce(PR_NUMBER.ADD_SVG);
+    inputs.getConfigFilePath.mockReturnValueOnce("this file doesn't exist");
 
     await main();