+++ CHANGELOG.md
@@ -9,6 +9,12 @@ Versioning].
 
 - _No changes yet_
 
+## [0.2.1] - 2020-03-02
+
+- Disable the action from a commit message. ([#91])
+- Add support for a separate configuration file. ([#92])
+- Improve the default commit message. ([#97])
+
 ## [0.2.0] - 2020-03-01
 
 - Commit all optimization in a single commit. ([#84])
@@ -44,3 +50,6 @@ Versioning].
 [#78]: https://github.com/ericcornelissen/svgo-action/pull/78
 [#81]: https://github.com/ericcornelissen/svgo-action/pull/81
 [#84]: https://github.com/ericcornelissen/svgo-action/pull/84
+[#91]: https://github.com/ericcornelissen/svgo-action/pull/91
+[#92]: https://github.com/ericcornelissen/svgo-action/pull/92
+[#97]: https://github.com/ericcornelissen/svgo-action/pull/97